net, sriov: Move helpers to new libsriov
Fetch mac from cloud init and remove duplicate
mac assignment
